feat: 상품 관련 api 비즈니스 로직, dto 추가

+++ src/main/java/kr/hhplus/be/server/presentation/dto/ProductTopResponse.java
@@ -2,16 +2,18 @@ package kr.hhplus.be.server.presentation.dto;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class ProductTopResponse {
 
     private final Integer rank;
     private final Long productId;
     private final String productName;
-    private final Long price;
+    private final BigDecimal price;
     private final Integer stock;
 
-    public ProductTopResponse(Integer rank, Long productId, String productName, Long price, Integer stock) {
+    public ProductTopResponse(Integer rank, Long productId, String productName, BigDecimal price, Integer stock) {
         this.rank = rank;
         this.productId = productId;
         this.productName = productName;
